### 2018-10-26 #1052 @whedon
[PRE REVIEW]: CITapp - a free reaction time-based Concealed Information Test application
**Submitting author:** @gasparl (<a href="http://orcid.org/0000-0001-9401-4830">Gáspár Lukács</a>)
**Repository:** <a href="https://github.com/gasparl/citapp_pc" target ="_blank">https://github.com/gasparl/citapp_pc</a>
**Version:** v1.0.0
