### 2018-10-26 #1052 @arfon
@gasparl - @whedon can't compile this paper as the `paper.md` and `paper.bib` don't seem to be valid UTF-8 . Could you try saving them as UTF-8 encoded and updating your repository?
