+++ paper/paper.bib
@@ -1,15 +1,15 @@
 @article{lukacs2017,
-  author = {Luk·cs, G., Kleinberg, B., and Verschuere, B.},
+  author = {Luk√°cs, G., Kleinberg, B., and Verschuere, B.},
   title = {Familiarity-related fillers improve the validity of reaction time-based memory detection},
   journal = {Journal of Applied Research in Memory and Cognition},
   volume = {6},
   issue = {3},
   year = {2017},
-  pages = {295ñ305},
+  pages = {295‚Äì305},
   doi = {10.1016/j.jarmac.2017.01.013}
 }
 @article{lukacs2018,
-  author = {Luk·cs, G. and Ansorge, U.},
+  author = {Luk√°cs, G. and Ansorge, U.},
   title = {The Adverse Effect of Information Leakage on the Response Time-Based Concealed Information Test with Inducers, Remedied by Omitting Targets. Manuscript submitted},
   year = {2018}
 }
