+++ citapp_documentation.md
@@ -2,7 +2,7 @@
 
 October 25, 2018
 
-Author: Gáspár Lukács
+Author: GÃ¡spÃ¡r LukÃ¡cs
 
 Contact: gaspar.lukacs@univie.ac.at
 
@@ -14,15 +14,15 @@ This application needs no installation whatsoever, it can and should be run in a
 
 **Introduction**
 
-No deception detection method has been scientifically proven to achieve perfect or close to perfect accuracies in discriminating deceptive ("guilty") persons from truth-telling ("innocent") ones, notwithstanding the claims by some commercial services and products. The most efficient "Concealed Information Test" version in the present application ("enhanced") achieved a remarkably high accuracy in a recent study (_Lukács, Kleinberg, & Verschuere, 2017_), but even so it can only be said to provide an accuracy of roughly around 90% – under certain conditions – and it still needs further research and replications. Correspondingly, **this application is only for testing and research purposes, and no serious judgments should be based on its results**. Furthermore, anyone using it should have at least a basic understanding of how it works – see below.
+No deception detection method has been scientifically proven to achieve perfect or close to perfect accuracies in discriminating deceptive ("guilty") persons from truth-telling ("innocent") ones, notwithstanding the claims by some commercial services and products. The most efficient "Concealed Information Test" version in the present application ("enhanced") achieved a remarkably high accuracy in a recent study (_LukÃ¡cs, Kleinberg, & Verschuere, 2017_), but even so it can only be said to provide an accuracy of roughly around 90% â€“ under certain conditions â€“ and it still needs further research and replications. Correspondingly, **this application is only for testing and research purposes, and no serious judgments should be based on its results**. Furthermore, anyone using it should have at least a basic understanding of how it works â€“ see below.
 
 **The CIT versions**
 
-This application implements two versions of the reaction time-based Concealed Information Test (RT-CIT; _Seymour et al., 2000_; _Lukács, Kleinberg, & Verschuere, 2017_): _standard_ and _enhanced_. Both of these aim to reveal whether or not a certain information detail – the _probe_ – is known to the tested person. **The identity of the probe should not be shown to the person taking the test: the starting page must be first completed and submitted by the person who administers the test. The**  **irrelevant**  **and**  **target** **items must be similar to the probe, and indistinguishable for a person who does not know the relevance of the probe (e.g. the probe is a stolen suitcase, and the irrelevant and targets are other suitcases).**
+This application implements two versions of the reaction time-based Concealed Information Test (RT-CIT; _Seymour et al., 2000_; _LukÃ¡cs, Kleinberg, & Verschuere, 2017_): _standard_ and _enhanced_. Both of these aim to reveal whether or not a certain information detail â€“ the _probe_ â€“ is known to the tested person. **The identity of the probe should not be shown to the person taking the test: the starting page must be first completed and submitted by the person who administers the test. The**  **irrelevant**  **and**  **target** **items must be similar to the probe, and indistinguishable for a person who does not know the relevance of the probe (e.g. the probe is a stolen suitcase, and the irrelevant and targets are other suitcases).**
 
 Standard RT-CIT: it has one probe, one target, and four irrelevant items. The target item will require a different response in the task, otherwise it is the same as an irrelevant item. A person who knows the relevance of the probe (e.g., because he stole the suitcase) will tend to make slower responses to it than to the irrelevant items.
 
[…]
