### 2018-10-26 #1052 @gasparl
Sure - I converted to UTF-8 and updated. Hope it's fine now.
