+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{lukacs2017,
-  author = {Lukács, G., Kleinberg, B., and Verschuere, B.},
+  author = {{Lukács}, G. and {Kleinberg}, B. and {Verschuere}, B.},
   title = {Familiarity-related fillers improve the validity of reaction time-based memory detection},
   journal = {Journal of Applied Research in Memory and Cognition},
   volume = {6},
@@ -9,12 +9,12 @@
   doi = {10.1016/j.jarmac.2017.01.013}
 }
 @article{lukacs2018,
-  author = {Lukács, G. and Ansorge, U.},
+  author = {{Lukács}, G. and {Ansorge}, U.},
   title = {The Adverse Effect of Information Leakage on the Response Time-Based Concealed Information Test with Inducers, Remedied by Omitting Targets. Manuscript submitted},
   year = {2018}
 }
 @article{seymour2000,
-  author = {Seymour, T. L., Seifert, C. M., Shafto, M. G., & Mosmann, A. L.},
+  author = {{Seymour}, T.~L. and {Seifert}, C.~M. and {Shafto}, M.~G. and {Mosmann}, A.~L.},
   title = {Using response time measures to assess "guilty knowledge"},
   journal = {Journal of Applied Psychology},
   volume = {85},
