### 2018-10-26 #1052 @gasparl
There were some small syntax mistakes in the paper.bib; I now corrected.
