### 2018-11-18 #1052 @yochannah
:wave: Hey @gasparl! I'll be editing this paper. Thanks for the list of suggested reviewers. While we're settling who will review the paper, it might be a good idea to look through [another review issue in this repo](https://github.com/openjournals/joss-reviews/labels/review) to get an idea of the things reviewers will be looking for. I'd recommend starting by adding a README.md to your bitbucket repo to provide basic guidelines for anyone visiting the repo.
