### 2018-11-18 #1052 @andytwoods
Kind of wish there was a readme and not just the pdf. And can't access the pdf on my fone as I'd need to download whole project! Happy to review if they add a readme.
