### 2018-11-18 #1052 @gasparl
Thanks very much @yochannah and @andytwoods! I now renamed the documentation md file to README.md - if these should be two separate texts (e.g., shorter README.md), do let me know.

I checked a couple of previous reviews already, but I will do another round very soon.
