### 2018-11-20 a3e7712
app testing function
