+++ README.md
@@ -1,6 +1,6 @@
 **Concealed Information Test application**
 
-October 25, 2018
+November 20, 2018
 
 Author: Gáspár Lukács
 
@@ -90,6 +90,14 @@ Still, for an immediate informative feedback for each test in this application, 
 | _d-CIT_ <=- 0.1 | fair indication of innocence |
 
 (Note that there is no strong indication of innocence: a very fast probe response is no more expected in case of innocence than in case of guilt.)
+
+**Testing and example usage**
+
+To experimentally test this application, one may for example enter the personal name (e.g., forename) of a participant as probe, and other, similar names as target and irrelevant items. (These latter items should be controlled for similarity to the extent possible; e.g., number of characters or syllables, word frequency, etc.) In this case, the participant's responses to the probe (own personal name) should be notably slower than the responses to the irrelevants (e.g., around 15-30 ms average difference). Correspondingly, the _d-CIT_ should be relatively high as well (e.g., larger than 0.3). As a control condition, one may enter irrelevant names for all item types, so that the probe, irrelevant, and target items are all in fact irrelevant to the participant. In this case, the responses to the probe should be of similar speed as the responses to the irrelevants (since, again, the probe is not relevant to the participant and is not recognized as any different from the rest of the items).
+
+However, the basic framework of the application can be tested programmatically as well. This is done by a designated testing function that simulates human responses (keypresses) whenever an item in the task is displayed. This simulation can be activated by simply entering the function _citapp\_testing\_on()_ into the Chrome DevTools Console (F12) anytime. By default, this function gives 100% correct responses in the first practice round (where this is required), and, for the rest of the task, 90% correct and 10% incorrect responses, with artificial response time randomly ranging from 350 to 600. These numbers can be changed by modifying any of the following parameters: _apptest\_correct\_ratio\_first_ (ratio from 0.0 to 1.0), _apptest\_correct\_ratio_ (ratio from 0.0 to 1.0), _apptest\_simulated\_rt\_min_ (number in ms), and _apptest\_simulated\_rt\_max_ (number in ms). For example, entering _apptest\_correct\_ratio\_first = 0.95, apptest\_correct\_ratio = 0.8, apptest\_simulated\_rt\_min = 200, apptest\_simulated\_rt\_max = 850_ into the Console will simulate a participant who has 95% correct responses in the first practice round, and, for the rest of the task, 80% correct and 20% incorrect responses, with artificial response time randomly ranging from 200 to 850 (hence, some of them over the response time limit of 800, resulting in "too slow" responses). The simulation automatically proceeds to the next block (within a few seconds) as long as there are blocks left in the task. This can be changed using the parameter _move\_to\_next\_block_ (true or false), e.g. _apptest\_move\_to\_next\_block = false_ will disable this, and the user has to press the appropriate button to start each block. By default, all items have the same simulated responses. However, guilt may be simulated using the variable _probe\_delay_ (number in ms), that adds a specified delay whenever the response is given to the probe. E.g., _apptest\_probe\_delay = 30_ will always add 30 ms to the probe response. To stop testing (i.e., stop simulating keypresses), enter _citapp\_testing\_off()_ anytime.
+
[…]
