### 2018-11-20 bced7ab
paper small corrections
