+++ paper/paper.bib
@@ -10,7 +10,7 @@
 }
 @article{lukacs2018,
   author = {{Lukács}, G. and {Ansorge}, U.},
-  title = {The Adverse Effect of Information Leakage on the Response Time-Based Concealed Information Test with Inducers, Remedied by Omitting Targets. Manuscript submitted},
+  title = {Information leakage in the response time-based Concealed Information Test. Manuscript submitted},
   year = {2018}
 }
 @article{seymour2000,
