### 2018-11-20 #1052 @gasparl
I went through the Review checklist once more, and I think everything should be fine. I now added a more thoroughly explained example usage, two example output files, and also an automatic testing function. I correspondingly updated the manual with the related information.
I also made some very small final corrections in the paper.
