### 2019-01-15 #1052 @yochannah
Quick update: I've had a couple more polite declines unfortuantely, mostly due to people being busy. I have another few leads to try though. Thank you for your patience!
