### 2019-01-15 #1052 @gasparl
Thanks @vsoch!
@yochannah: should I move it to Github? It's totally fine by me.
