### 2019-01-15 #1052 @vsoch
I highly encourage moving to Github :) At quick glance I see something that I would really enjoy (after review) adding static Github pages preview / containers with continuous integration, and (it's probably doable on bitbucket) but I like Github much better. At least every time I've tried bitbucket I can't ever find what I'm looking for, lol.
