### 2019-01-15 #1052 @gasparl
It's really no problem, I've been planning to gradually move my all code to Github anyway.
Here is the new repository: https://github.com/gasparl/citapp_pc
I think it should be all there and fine - but if there is anything else I should do, let me know.
