### 2019-01-15 #1179 @whedon
[REVIEW]: CITapp - a free reaction time-based Concealed Information Test application
**Submitting author:** @gasparl (<a href="http://orcid.org/0000-0001-9401-4830">Gáspár Lukács</a>)
**Repository:** <a href="https://github.com/gasparl/citapp_pc" target ="_blank">https://github.com/gasparl/citapp_pc</a>
**Version:** 1.0.3
