### 2019-01-15 #1179 @vsoch
Note that I'm updating the current bitbucket repo in whedon's message to be https://github.com/gasparl/citapp_pc
