### 2019-01-15 gasparl/citapp_pc#2 @vsoch
2. Summary opening assumes previous knowledge
You open with:

> This application implements two versions of the response time-based Concealed Informa-
tion Test (RT-CIT) deception detection method: a standard version that has been often
used in the recent years (Seymour, Seifert, Shafto, & Mosmann, 2000), and a recently
introduced enhanced version (Lukács, Kleinberg, & Verschuere, 2017)

referencing "the response time-based Concealed Information Test" as if I know what you
are talking about. I have no idea. I would suggest to rewrite the summary for the user
that doesn't know what you are talking about. For example, something along an outline like:

 1. It's important to study how people recognize items (because...)
 2. Thus they developed this task called the Concealed.... (references)
 3. The task is great, but there are still pressing problems that (it's not reproducible, hard to run, etc.)
 4. Thus, to solve these issues, we created this! :)

I would rewrite the summary in that context - you can still provide some of the more
"academic paper" sounding details of the task, but remember that the paper here is
intended for a general audience, and not a niche psychology community.

Review: https://github.com/openjournals/joss-reviews/issues/1179
