### 2019-01-15 gasparl/citapp_pc#3 @vsoch
3. Online version
You reference an "online version" on osf, but this isn't easy enough to use or interact
with. I'd recommend you deploy the static files to Github pages for a preview
that lives with (and thus is versioned with) the codebase. I would recommend having
the demo in a "docs" folder on the master branch so you don't need to worry about updating
a separate branch (gh-pages). If you need any help / have questions please ask and I will
help you!

Review: https://github.com/openjournals/joss-reviews/issues/1179
