### 2019-01-15 gasparl/citapp_pc#4 @vsoch
4. User interface
It would be useful to have a figure that shows the user interface in the paper.

Review: https://github.com/openjournals/joss-reviews/issues/1179
