### 2019-01-15 gasparl/citapp_pc#5 @vsoch
5. Repository description and tags
Make sure to add tags and a description to your repository so it's better indexed on Github

Review: https://github.com/openjournals/joss-reviews/issues/1179
