### 2019-01-15 gasparl/citapp_pc#6 @vsoch
6. Instructions for contribution and getting support
In the future, it could be that someone wants to contribute to your code! The repository is missing instructions in the README.md to do this, and it's also commonly seen to have a .github folder with a CONTRIBUTING.md file. You can even find [templates](https://gist.github.com/PurpleBooth/b24679402957c63ec426) if you do a Google search. The example that I linked also includes a Code of Conduct, and this is also important to add. You can also add it as another file in .github, CODE_OF_CONDUCT.md. You always want to have a standard for interaction in the case that someone is behaving inappropriately in an issue or PR (it happens, if you can't believe it!)

It's also likely that someone will want to ask you a question or get help. For this, you can just have a section in the readme (Support) and then link them to your issues board here.

Review: https://github.com/openjournals/joss-reviews/issues/1179
