### 2019-01-15 gasparl/citapp_pc#7 @vsoch
7. "Installation" or Usage Instructions
The installation, based on the README, is to "drag and drop the file in a browser."
As a stupid Github user I could try grabbing the link to the file and dropping it in a new
tab, and this obviously wouldn't work! I would suggest a more direct instruction that
can first direct the user to a static url to preview the experiment, and then
if the user wants to deploy  on his / her own, provide complete instructions to:

 1. Fork the repository to their Github account
 2. Clone the forked repository
 3. Customize (in some way) if necessary?
 4. Either put on a local web server, or give instructions to deploy to Github pages.

Also, you could provide the user with a Dockerized version, which would let them
run a Docker container bound to port 80 to run it immediately without needing to do
cloning, etc. This would also be reproducible because you would have all versions
of the software build into containers. let me know if you need help with this, glad
to help with creating a Dockerfile and building the container with some CI (e.g., circleci)
and pushing the container to Docker Hub.

Review: https://github.com/openjournals/joss-reviews/issues/1179
