### 2019-01-16 gasparl/citapp_pc#3 @gasparl
Brilliant, I didn't know it's so easy to host my page on Github - thank you for the suggestion!
I'm not sure though what you mean by "having
the demo in a "docs" folder". The application is now simply acceptable via https://gasparl.github.io/citapp_pc/CITapp.html
I would just leave it like that and refer to this in the repository and in the paper.
