+++ paper/paper.bib
@@ -8,10 +8,10 @@
   pages = {295–305},
   doi = {10.1016/j.jarmac.2017.01.013}
 }
-@article{lukacs2018,
+@article{lukacs2019,
   author = {{Lukács}, G. and {Ansorge}, U.},
   title = {Information leakage in the response time-based Concealed Information Test. Manuscript submitted},
-  year = {2018}
+  year = {2019}
 }
 @article{seymour2000,
   author = {{Seymour}, T.~L. and {Seifert}, C.~M. and {Shafto}, M.~G. and {Mosmann}, A.~L.},
