+++ README.md
@@ -32,7 +32,7 @@ _These inducer items induce a particular semantic context that increases the sen
 
 The number of blocks can be chosen (1, 2, or 3). One block has the same length and arrangement as in all recent studies (_Lukács, Kleinberg, & Verschuere, 2017_). There is no data to show how many blocks per item categories (one probe with corresponding target and irrelevants) are optimal, but, as a rule of thumb, one block should give a fair estimation on the group level, while more than three would probably not serve much additional information even on the individual level.
 
-_In each block, each probe, irrelevant, and target is repeated 18 times (hence altogether 54 probe, 216 irrelevant, and 54 target trials). The order of the items is randomized in groups: first, all six items (one probe, four irrelevant, and one target) in the given category are presented in a random order, then the same six items are presented in another random order (but with the restriction that the first item in the next group is never the same as the last item in the previous group). In the enhanced version, inducers are placed among these items in a random order, but with the restrictions that an inducer trial is never followed by another inducer trial, and each of the nine inducers (three familiarity-referring, six unfamiliarity- referring) precedes each of the three probes, three targets (for Induced, but not for No-target), and 12 irrelevants exactly one time. (Thus 9 × 18 = 162 inducers are presented in the Induced CIT, and 162 out of the 324 other items are preceded by an inducer.)_
+_In each block, each probe, irrelevant, and target is repeated 18 times (hence 18 probe, 72 irrelevant, and 18 target trials). The order of the items is randomized in groups: first, all six items (one probe, four irrelevant, and one target) in the given category are presented in a random order, then the same six items are presented in another random order (but with the restriction that the first item in the next group is never the same as the last item in the previous group). In the enhanced version, inducers are placed among these items in a random order, but with the restrictions that an inducer trial is never followed by another inducer trial, and each of the nine inducers (three familiarity-referring, six unfamiliarity- referring) precedes each of the probes, targets, and four irrelevants exactly one time. (Thus 9 × 6 = 54 inducers are presented in the Induced CIT, and 54 out of the 108 other items are preceded by an inducer.)_
 
 _The main task is preceded by a comprehension check and two practice tasks. The check serves to ensure that the participant fully understands the task. This phase includes all items presented once, including each of the nine possible inducers for the enhanced version. In this phase, participants has plenty of time (10 s) to choose a response – however, each trial requires a correct response. In case of an incorrect response, the phase has to be repeated._
 
