### 2019-01-16 gasparl/citapp_pc#3 @vsoch
For the demo, it might be helpful to add hints / pre-filled in values for the content on the right (I don't know what to put) and most that are previewing likely won't have read through any documentation yet.

![image](https://user-images.githubusercontent.com/814322/51259240-bb8fa000-1979-11e9-9c16-c2781abc186e.png)

This would be the kind of reason to have the demo separate from the thing you are giving to users to deploy - the demo would have more learning tips / pre-filled in boxes for teaching.
