+++ CONTRIBUTING.md
@@ -1,0 +1,37 @@
+# Contributing
+
+Contributions are always welcome. When contributing to this repository, please first discuss the change you wish to make via issue, email, or any other method with the owner of this repository before making a change. 
+
+## All Code Changes Happen Through Pull Requests
+
+1. Fork the repo and create your branch from `master`.
+2. If you've added code that should be tested, add tests.
+3. If you've changed any relevant functionalities, update the README.md.
+4. Ensure the application is working correctly (e.g. using the automatic testing functions).
+5. Make sure the syle of your code is consistent with that of the current one (indentation, etc.).
+6. Issue that pull request.
+
+## Code of Conduct
+
+Examples of behavior that contributes to creating a positive environment
+include:
+
+* Using welcoming and inclusive language
+* Being respectful of differing viewpoints and experiences
+* Gracefully accepting constructive criticism
+* Focusing on what is best for the community
+* Showing empathy towards other community members
+
+Examples of unacceptable behavior by participants include:
+
+* Trolling, insulting/derogatory comments, and personal or political attacks
+* Public or private harassment
+* The use of sexualized language or imagery and unwelcome sexual attention or
+advances
+* Publishing others' private information, such as a physical or electronic
+  address, without explicit permission
+* Other conduct which could reasonably be considered inappropriate in a
+  professional setting
+  
+## License
+By contributing, you agree that your contributions will be licensed under its BSD-2-Clause License.
