### 2019-01-17 2e1549b
documentation update
