+++ paper/paper.md
@@ -13,21 +13,25 @@ authors:
 affiliations:
  - name: University of Vienna, Department of Basic Psychological Research and Research Methods, Austria
    index: 1
-date: 16 January 2019
+date: 17 January 2019
 bibliography: paper.bib
 ---
 
 # Summary
 
-This application implements two versions of the response time-based Concealed Information Test (RT-CIT) deception detection method: a standard version that has been often used in the recent years [@seymour2000], and a recently introduced enhanced version [@lukacs2017]. Both of these aim to reveal whether or not a certain information detail – the probe – is known to the tested person. Along with this probe, the test also contains irrelevant items that are similar to the probe, and thus indistinguishable for a person who does not know the relevance of the probe (e.g., when the probe is the brand of a stolen car, the irrelevant items are other brands). When these items are repeatedly presented in a test, a persons who recognize the relevance of the probe (e.g., because they were the ones who stole the car) will tend to make slower keypress responses to the probe than to the irrelevant items. Consequently, based on these larger probe-irrelevant response time differences, these persons can be distinguished from those who do not know the relevance of the probe (e.g., because they took no part in the stealing of the car).
+Undetected deception may have extremely high costs in certain scenarios, such as counterterrorism, pre-employment screening for intelligence agencies, or high-stake criminal proceedings. However, meta-analyses have repeatedly shown that without special aid, based on their own best judgment only, people (including police officers, detectives, and professional judges) distinguish lies from truth on a level hardly better than mere chance (e.g., [@hartwig2011]). 
 
-While the standard version is simple in design, and therefore relatively easy to implement, the recent enhanced version [@lukacs2017] is much more complex (especially in respect of its item randomization). Therefore, the primary aim of this application is to provide a readily available tool for the independent testing and replication of this method. This application requires no technical knowledge; it can simply be opened and used in a Google Chrome browser. Furthermore, by simply uploading it on any server, it can also be made available online, where it can then be easily accessed anytime, from anywhere. One online version is always available directly via the [site of the repository itself](https://gasparl.github.io/citapp_pc/CITapp.html "CITapp"). Once loaded into the browser, it may again be used offline as well.
+The response time-based Concealed Information Test (RT-CIT) aims to reveal whether a person is concealing knowledge regarding a certain detail [@seymour2000]. To illustrate the CIT, let us consider a murder case scenario in which the murder weapon is known only to the perpetrator and the investigators. In this case, the CIT could include the actual murder weapon (the probe; e.g. "rifle") and several other weapons (irrelevants; e.g. "knife," and "rope") as the items that would be sequentially presented to a suspect in a random order. When an examinee needs to respond to each item with a keypress, the recognition of the probe (in this case, "rifle") by a guilty person (who is aware of the relevance of that item) will typically result in a slower response to that item than to the irrelevant items. Thereby, based on the probe-irrelevant RT differences, a guilty person can be distinguished from innocent ones.
 
-The startpage looks like this:
+The standard RT-CIT includes a single (randomly chosen) target irrelevant item that requires pressing a response key different from the response key for probe and other, nontarget irrelevant items. For example, the key "I" has to be pressed whenever the target item appears, while the key "E" has to be pressed whenever any of the other items (probe and nontarget irrelevants) appear. This target item, which appears at random times, ensures that the examinees keep paying attention to the changing items. However, a recent study improved this method by adding filler items to the task. These were hypothesized to induce even closer attention to the items, and hence also to the probe [@lukacs2017]. This method robustly increased the CIT accuracy of distinguishing guilty examinees from innocent ones. This enhanced version, however, is much more difficult to implement, especially due to the complex randomization rules in respect of the filler items. Therefore, the primary aim of this application is to provide a readily available tool for the independent testing and replication of this method.
 
[…]
