+++ paper/paper.bib
@@ -23,3 +23,13 @@
   pages = {30-37},
   doi = {10.1037//0021-9010.85.1.30}
 }
+@article{hartwig2011,
+  author = {{Hartwig}, M. and {Bond}, C.~F.},
+  title = {Why do lie-catchers fail? A lens model meta-analysis of human lie judgments},
+  journal = {Psychological Bulletin},
+  volume = {137},
+  issue = {4},
+  year = {2011},
+  pages = {643–659},
+  doi = {10.1037/a0023589}
+}
