### 2019-01-19 b0629c4
documentation update
