### 2019-01-19 gasparl/citapp_pc#9 @andytwoods
[REVIEW] indepth
Hi,
well done writing this. Hard to achieve, this sort of thing.
Some comments:
-in the readme, the pic links are broken and are not shown.
-software engineers view 'automated tests' quite differently from your 'simulated human' tests (although your simulated human tests were very worthwhile developing). The engineer would use software (e.g. travis?) to run a batch of very specific tests to run upon each git push, to ensure things are not broken. These specific tests typically would test their code on a function by function basis (unit testing) and in bigger chunks of functionality (many functions). I personally use Tape when testing JS (there a many testing frameworks though; https://medium.com/javascript-scene/why-i-use-tape-instead-of-mocha-so-should-you-6aa105d8eaf4). I would advise creating some tests for the main functions in your code for peace of mind, and for future proofing.
