### 2019-01-19 #1179 @andytwoods
Sorry for delays. Started a new job and things still settling down. A nice project. My minimal comments: https://github.com/gasparl/citapp_pc/issues/9
