+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'CITapp - a lie detector web application for the response time-based Concealed Information Test'
+title: 'CITapp - a response time-based Concealed Information Test lie detector web application'
 tags:
   - concealed information test
   - deception detection
@@ -21,9 +21,9 @@ bibliography: paper.bib
 
 Undetected deception may have extremely high costs in certain scenarios, such as counterterrorism, pre-employment screening for intelligence agencies, or high-stake criminal proceedings. However, meta-analyses have repeatedly shown that without special aid, based on their own best judgment only, people (including police officers, detectives, and professional judges) distinguish lies from truth on a level hardly better than mere chance (e.g., [@hartwig2011]). 
 
-The response time-based Concealed Information Test (RT-CIT) aims to reveal whether a person is concealing knowledge regarding a certain detail [@seymour2000]. To illustrate the CIT, let us consider a murder case scenario in which the murder weapon is known only to the perpetrator and the investigators. In this case, the CIT could include the actual murder weapon (the probe; e.g. "rifle") and several other weapons (irrelevants; e.g. "knife," and "rope") as the items that would be sequentially presented to a suspect in a random order. When an examinee needs to respond to each item with a keypress, the recognition of the probe (in this case, "rifle") by a guilty person (who is aware of the relevance of that item) will typically result in a slower response to that item than to the irrelevant items. Thereby, based on the probe-irrelevant RT differences, a guilty person can be distinguished from innocent ones.
+The response time-based Concealed Information Test (RT-CIT) aims to reveal whether a person is concealing knowledge regarding a certain detail [@seymour2000]. To illustrate the CIT, let us consider a murder case scenario in which the murder weapon is known only to the perpetrator and the investigators. In this case, the CIT could include the actual murder weapon (the probe; e.g. "rifle") and several other weapons (irrelevants; e.g. "knife," and "rope") as the items that would be sequentially presented to a suspect in a random order. When each item has to be responded to with a keypress, the recognition of the probe (in this case, "rifle") by a guilty person (who is aware of the relevance of that item) will typically result in a slower response to that item than to the irrelevant items. Thereby, based on the probe-irrelevant RT differences, a guilty person can be distinguished from innocent ones.
 
-The standard RT-CIT includes a single (randomly chosen) target irrelevant item that requires pressing a response key different from the response key for probe and other, nontarget irrelevant items. For example, the key "I" has to be pressed whenever the target item appears, while the key "E" has to be pressed whenever any of the other items (probe and nontarget irrelevants) appear. This target item, which appears at random times, ensures that the examinees keep paying attention to the changing items. However, a recent study improved this method by adding filler items to the task. These were hypothesized to induce even closer attention to the items, and hence also to the probe [@lukacs2017]. This method robustly increased the CIT accuracy of distinguishing guilty examinees from innocent ones. This enhanced version, however, is much more difficult to implement, especially due to the complex randomization rules in respect of the filler items. Therefore, the primary aim of this application is to provide a readily available tool for the independent testing and replication of this method.
[…]
