+++ paper/paper.md
@@ -25,7 +25,7 @@ The response time-based Concealed Information Test (RT-CIT) aims to reveal wheth
 
 The standard RT-CIT includes a single (randomly chosen) target irrelevant item that requires pressing a response key different from the response key for probe and other, nontarget irrelevant items. For example, the key "I" has to be pressed whenever the target item appears, while the key "E" has to be pressed whenever any of the other items (probe and nontarget irrelevants) appear. This target item, which appears at random times, ensures that the examinees keep paying attention to the changing items. However, a recent study improved this method by including additional filler items that, similarly as the target, appear randomly throughout the task and have to be categorized with one of the two response keys. These were hypothesized to induce even closer attention to the items, and hence also to the probe [@lukacs2017]. This method robustly increased the CIT accuracy in distinguishing guilty examinees from innocent ones. This enhanced version, however, is much more difficult to implement, especially due to the complex randomization rules in respect of the filler items. Therefore, the primary aim of this application is to provide a readily available tool for the independent testing and replication of this method.
 
-This application requires no technical knowledge; it can simply be opened and used in a Google Chrome browser. Furthermore, by uploading it on any server, it can also be made available online, where it can then be easily accessed anytime, from anywhere. One online version is always available directly via the [site of the repository itself](https://gasparl.github.io/citapp_pc/CITapp.html "CITapp"). Once loaded into the browser, it may again be used offline as well.
+This application requires no technical knowledge; it can simply be opened and used in a Google Chrome browser. Furthermore, by uploading it on any server, it can also be made available online, where it can then be easily accessed anytime, from anywhere. One online version is always available directly via the site of the repository itself: https://gasparl.github.io/citapp_pc/CITapp.html. Once loaded into the browser, it may again be used offline as well.
 
 The details and settings of the test can be entered and selected, including the specification of the probe, irrelevant, and target items, on the start page; see Figure 1. After these basic information details are given, the test may be initiated. The examinee then simply has to follow the instructions on the screen. The full results data (and also a preliminary evaluation) are automatically provided and downloadable after the completion of the test.
 
