### 2019-01-20 gasparl/citapp_pc#1 @gasparl
I renamed it as "CITapp - a response time-based Concealed Information Test lie detector web application".
(Even if not the best from the layman perspective, I would keep the "Concealed Information Test" part in the title, because it is simply the proper name of the method, already used in hundreds of publications.)
