### 2019-01-20 gasparl/citapp_pc#2 @gasparl
I revised the paper along the lines of your suggestion; mainly more detailed introduction and a screenshot of the GUI.
