### 2019-01-20 gasparl/citapp_pc#3 @gasparl
I added a demo possibility that can automatically fill these example using a button, but, for compactness, I simply added it as a new menu in the main application:
![image](https://user-images.githubusercontent.com/37374850/51437582-88624f00-1ca0-11e9-9575-bea083cdf1d8.png)

So now it's available via the same page; https://gasparl.github.io/citapp_pc/CITapp.html. I added this link both to the paper and the README.
