### 2019-01-20 gasparl/citapp_pc#6 @gasparl
Yes, I missed this, again thanks for noting. I now added the CONTRIBUTING.md file (with a short Code of Conduct inside it).

I also added the Support section in the README.
