### 2019-01-20 gasparl/citapp_pc#7 @gasparl
I added more precise instructions about how the application may be used.

I'm not sure though about the Dockerized version. I am completely new to this, but if I understand correctly the only advantage would be that separate versions are kept unchanged in these containers. But then downloading or cloning could do just the same, and these can be done extremely easily, literally with a few button clicks. (The huge green "Clone or download" button for the repository.)
But if I misunderstand and there is more to this and/or you really think it would be relevant, then I would indeed appreciate it you helped out with the implementation :)
