### 2019-01-20 gasparl/citapp_pc#8 @gasparl
Thank you for your thorough review. Here I just reply to those few things that were not mentioned in separate issues.

- I renamed the LICENSE.txt to LICENSE. (It was just an oversight.)

- I added several screenshots to the README.

- I noted the availability (and link) of the PDF version in the README.
