### 2019-01-20 gasparl/citapp_pc#3 @vsoch
This is totally awesome! I both love it and can say you went above my expectation to make it easier to learn how to use. This definitely closes this issue. Great work!
