### 2019-01-20 gasparl/citapp_pc#8 @vsoch
Great work @gasparl . I had a few outstanding questions in the other issues, but I'm happy with the changes once we get those closed up (and I'll make final edits to the joss-reviews issue and finish checking off boxes).
