### 2019-01-20 gasparl/citapp_pc#2 @gasparl
Thank you!

As you may have seen there are several warning messages on the start page whenever any browser other than Chrome is used. In my related online experiments I did not only put warnings but actually completely disabled the task for other browsers. Always using the same browser ensures that everything works and is displayed exactly the same way; so that it would be consistent both within and across experiments/studies/individual tests. Since this is not some website for general visits but an experimental tool, this doesn't seem much of a practical limitation. (That is: for someone conducting an experiment, downloading Chrome - if not already in use - is the least of the trouble.)
I confess though that I decided on this mainly because of the time it took to each time test and make the code compatible with all browsers, which was especially and repeatedly problematic with IE. (Now I made a few quick tests; IE again doesn't work, Firefox and Opera seems fine.) Using one browser hugely simplifies this. But I also had some more nuanced issues, e.g., based on my tests, [synchronizing code execution to screen repaint](https://developer.mozilla.org/en-US/docs/Web/API/window/requestAnimationFrame) only seems to work well (for my timing purposes) on Chrome. In general, everything seems to work most easily and best on Chrome. And [everyone uses Chrome](https://en.wikipedia.org/wiki/File:Usage_Share_of_browsers_(updated_August_2018).png) anyway.
