### 2019-01-20 gasparl/citapp_pc#7 @gasparl
I see.
It's really nice of you to offer. Still, it doesn't seem pressing, so I would like to ask you to let me off the hook for now. If someone does ask for it and/or this test will actually become frequently used (as I would wish), I will not hesitate to get back to you about this.
