### 2019-01-20 gasparl/citapp_pc#2 @vsoch
I think it's okay that you are clear to specify the browser to use - it certainly is hard to maintain support for all of them. I was mainly curious and double checking that it wasn't an error to only specify Chrome.
