### 2019-01-20 #1179 @vsoch
@yochannah - an update from me! @gasparl has responded to and fixed [all of my issues](https://github.com/gasparl/citapp_pc/issues?q=is%3Aissue+is%3Aclosed), and done a fantastic job. Specifically, he has better written the manuscript for a general audience, and adding guidelines for contributing, getting help, and usage to the repository README, along with adding a demo depoyment to Github pages.

The only box unchecked is with respect to automated tests. It looks like @andytwoods has made some suggestions for how to go about this in his review, so when @gasparl has completed any remaining issues to the satisfaction of @andytwoods I give my thumbs up as well. Bravo, overall great job @gasparl !
