### 2019-01-20 #1179 @gasparl
Thanks very much @vsoch for your really kind and useful review!

(Yes, the testing function is in progress.)
