### 2019-01-21 #1179 @vsoch
I'm really interested to learn about the CI strategy that is underway to test the web interface. I've done some testing with selenium (via python) but I don't think that's best practice, so I'm looking forward to learning the details. :)
