### 2019-01-21 15ac8e0
dev_test function
