+++ CONTRIBUTING.md
@@ -6,9 +6,9 @@ Contributions are always welcome. When contributing to this repository, please f
 
 1. Fork the repo and create your branch from `master`.
 2. If you've added code that should be tested, add tests.
-3. If you've changed any relevant functionalities, update the README.md.
-4. Ensure the application is working correctly (e.g. using the automatic testing functions).
-5. Make sure the syle of your code is consistent with that of the current one (indentation, etc.).
+3. Make sure the syle of your code is consistent with that of the current one (indentation, etc.).
+4. If you've changed any relevant functionalities, update the README.md.
+5. Ensure the application is working correctly. The _dev\_test()_ function should pass without issues (see under _Testing and example usage_ in the [README.md](https://github.com/gasparl/citapp_pc/blob/master/README.md "README").
 6. Issue that pull request.
 
 ## Code of Conduct
