### 2019-01-21 90e231c
dev_test function
