### 2019-01-21 gasparl/citapp_pc#9 @gasparl
Hi, thank you for your review!

Regarding the images in the README, I just forgot to push them; now corrected.

Regarding the testing, I now wrote a function that does the following (quoted from the updated README):

" [...] executes all main functions that initiate the test, prepare the conditions, and, most importantly, generate the test items. The start page input fields are filled automatically using the _Demo_ data and settings. The function also checks whether, in case of either CIT version (standard or enhanced), the expected number of items are generated with the expected item texts (i.e., with the words that are entered on the start page). If no error occurs, the approval message is logged in the Console. Additionally, the function may be given a single argument: the number 1, or the number 2. Entering either will automatically initiate human simulation (with _apptest\_probe\_delay_ set to 25) after (and if) the previously described test was passed. Choosing _dev\_test(1)_ will complete the standard version, while _dev\_test(2)_ will complete the enhanced version. In case of no issues, the test should complete itself automatically, all the way until the end of test page. "

I also noted this test as a requirement in the CONTRIBUTING.md.

Now, I also checked out Tape, and I think I could implement this function in it quite straightforwardly - but I wouldn't think this elaboration (with its 2.4 MB extra weight) is necessary for this relatively simple and light application.

In any case, let me know if you are alright with this.
