+++ js/utils.js
@@ -20,12 +20,12 @@ function valid_nums(evt) {
 function fill_up_demo() {
     $("#exp_title_id").val("detect_relevant_date");
     $("#subj_num_id").val("suspect_01");
-    $("#probe_id").val('May 9');
-    $("#targ_id").val('August 25');
-    $("#irr1_id").val('June 14');
-    $("#irr2_id").val('December 5');
-    $("#irr3_id").val('February 12');
-    $("#irr4_id").val('March 29');
+    $("#probe_id").val('MAY 09');
+    $("#targ_id").val('AUG 25');
+    $("#irr1_id").val('JUN 14');
+    $("#irr2_id").val('DEC 05');
+    $("#irr3_id").val('FEB 12');
+    $("#irr4_id").val('MAR 29');
 }
 
 function valid_chars(evt) {
