### 2019-01-25 gasparl/citapp_pc#9 @vsoch
hey @gasparl ! I'm just peeking in on the review - the tests you added are an improvement, and I think what @andytwoods is looking for is automated testing, namely via continuous integration. This would mean that you wouldn't need to install / include a library like tape here, but you might have it set up with a service like CircleCI or Travis CI so your tests are run with each pull request / merge into master. Here are some links that might be helpful (personally I like CircleCI):

 - https://discuss.circleci.com/t/2-0-parallelism-with-nodejs-and-tape/22900/3
 - https://circleci.com/docs/2.0/browser-testing/
 - https://github.com/dwyl/learn-tape

I'm green to many of these testing suites (although I've done testing with selenium and python!) but I'd be happy to help (or try!) if you decide on a strategy that you like best. Minimally, you could create a CI configuration that opens a headless browser, issues the testing functions, and gets back some result.
