### 2019-01-25 gasparl/citapp_pc#9 @vsoch
It might seem like a hassle now, but I can assure you down the line it makes life so much easier! Because whenever you do a change, or someone opens a PR, all the tests are run automatically (without you needing to think) and you are alerted in the CI (directly in the pull request interface) if a new change has made something go awry.
