### 2019-01-26 gasparl/citapp_pc#9 @gasparl
The hassle is no problem, I'm just trying to keep things as simple as possible. I opened a question on SO for this: https://stackoverflow.com/questions/54372069/simplest-way-to-test-javascript-code-for-web-frontend-on-each-git-push
Either of you is welcome not only to answer/comment but also to edit the question if you wish.

If there is no good answer soon, I will probably just use Travis, that seems the most straightforward based on what I read so far.
