### 2019-01-26 gasparl/citapp_pc#9 @vsoch
I've heard good things about sauce labs too, they are free for open source! Some docs [here](https://docs.travis-ci.com/user/gui-and-headless-browsers/) and my colleague's repo where he uses it -> https://github.com/FelixHenninger/lab.js/blob/master/.travis.yml.
