### 2019-01-26 gasparl/citapp_pc#9 @vsoch
[CirceCI](https://circleci.com/docs/2.0/browser-testing/#sauce-labs)
