### 2019-01-26 gasparl/citapp_pc#9 @vsoch
hey @gasparl I'm [working on](https://github.com/researchapps/citapp_pc/blob/add/circleci/.circleci/test.py) helping to add a continuous integration set up to run the tests, and I've tested the dev_test(1) and dev_test(2) and I think this will work! What we need is some signal from the browser (that the headless tester robot of sorts) can check for to know if the tests are done (and then if they have failed). For example, I would want to be able to issue there commands to the browser:

```javascript
# start tests, maybe don't return anything
dev_test(1)

# then every 10 seconds, check test status
return tests_complete;

# if the above is True, we finish the script and everything passes / green. If not, we can ask for an additional variable (or run a function) to get the failing result.

completed = self.browser.execute_script("return tests_complete");
if completed:
     success = self.browser.execute_script("return tests_success");
     if not success:
          # and then we would want a function or variable to grab here to print for the user to see.
```
Does that make sense? This is a bit different than I think a normal testing scenario (that has different modular tests that are run individually) but since you have a cohesive, timed task, I think it would be okay to have it run in one long test as long as we are able to see the result, and if there is a failure, to see why. Also note that I'm using a chrome driver since that seemed to be an important detail.

Let me know what you would like to try!
