### 2019-01-26 gasparl/citapp_pc#9 @gasparl
@vsoch you are super nice helping out with this. How exactly does this work, how do I initiate the tests?

I don't think the dev_test(1) and dev_test(2) are necessary to be run on every change, I would just put dev_test() in case of `git push`. The former two is mainly for checking (by eye) that the display changes as expected. I can hardly imagine an automated test for that. Still, if you want I can of course add some value that is returned in case of a successfully completed test.

In general, let me know if I can add anything.

The 15 MB Chrome driver is not too great, but okay I can live with it.

By the way, I've been trying in a different direction, I just arrived at a working draft with Travis and Jasmin/Karma (https://github.com/gasparl/citapp_pc/tree/issue9), but it'd be still long way to go and I'd be happy to abandon it.
