### 2019-01-26 gasparl/citapp_pc#9 @vsoch
hey @gasparl ! If you have a method you are working on, then please continue! I'm not a professional front end developer so the most I can help do is running a headless browser with selenium, and having Python be the controller.

Basically, think of a headless browser starting up, and we can navigate to the experiment page. We can also feed the dev_test() into the browser, and get things running. The challenge, however, is getting the *result* of the tests back to the controller. There must be an additional command we can send to the browser (meaning equivalent to something we could type into the javascript console) to get a variable that tells us if tests are complete and passing. Since the testing takes some time, we would need to just check periodically. So this is what I'm asking - I need dev_test() to change some variables that I can check every once in a while to get an update, OR a separate function to run that returns some result that I can parse. Does that make sense? The tests are running headless so there is nothing to look at the browser, or determine that anything is being printed (unless we directly issue a command that can ask for a specific variable).
