+++ .travis.yml
@@ -3,3 +3,4 @@ node_js:
  - "stable"
 install:
   - npm install
+script: npm test
