+++ test/add_html.js
@@ -1,0 +1,5 @@
+var fixture = "<div id='div_intro_general' class='intro' style='text-align:left;padding: 0;'> <div style='padding: 3%;overflow: auto;height:78%;'> <div id='cit_tab'> <div style='display: flex;'> <div style='flex: 57%;line-height: 1.7;'> Experiment title: <input type='text' onkeypress='valid_chars(event)' id='exp_title_id' class='form_start' maxlength='20' size='20' value='CIT_app_test'> <br><br> Subject ID: <input type='text' onkeypress='valid_chars(event)' id='subj_num_id' class='form_start' maxlength='15' size='15'> <br><br> CIT version: <input type='radio' id='standard_id' name='cit_version' value='0'><label for='standard_id'>standard</label> <input type='radio' id='enhanced_id' name='cit_version' checked='checked' value='1'><label for='enhanced_id'>enhanced</label> <!--<input type='radio' id='notarget_id' name='cit_version' value='2'><label for='notarget_id'>no-target</label>--> <br><br> Number of blocks: <input type='radio' id='1block' name='num_of_blcks' checked='checked' value='1'><label for='1block'>1</label> <input type='radio' id='2block' name='num_of_blcks' value='2'><label for='2block'>2</label> <input type='radio' id='3block' name='num_of_blcks' value='3'> <label for='3block'>3</label> </div> <div style='flex: 43%;'> Probe: <input type='text' id='probe_id' class='form_start' maxlength='40' size='15'> <br> <br>Target: <input type='text' id='targ_id' class='form_start' maxlength='40' size='15'> <br> <br>Irrelevant (1): <input type='text' id='irr1_id' class='form_start' maxlength='40' size='15'> <br> <br>Irrelevant (2): <input type='text' id='irr2_id' class='form_start' maxlength='40' size='15'> <br> <br>Irrelevant (3): <input type='text' id='irr3_id' class='form_start' maxlength='40' size='15'> <br> <br>Irrelevant (4): <input type='text' id='irr4_id' class='form_start' maxlength='40' size='15'> <br> <br> </div> </div> <br> <i><b>Hints:</b> &#9679; close major running programs &#9679; disconnect internet &#9679; set fullscreen <span style='font-size:1.3vw;'>(<b>F11</b> or <b>Fn + F11</b> or <b>Ctrl + Cmd + F</b>)</span></i> <br> <div id='not_chrome_warn' style='display: none;color:red;'></div> <br> <hr> </div> <div id='settings_tab' style='display: none;'> <div style='display: flex;'> <div style='flex: 50%;'> <b><u>Reminder captions</u> (practice phase)</b> <br><br> <b>Standard version</b> <br><br> Caption at the top: <input type='text' id='s_top_id' class='form_start' maxlength='40' size='15' value='recognize?'> <br> <br>Caption at the left: <input type='text' id='s_left_id' class='form_start' maxlength='40' size='15' value='no = 'E''> <br> <br>Caption at the right: <input type='text' id='s_right_id' class='form_start' maxlength='40' size='15' value='yes = 'I''> <br> <br> <br> <b>Enhanced version</b> <br><br> Caption at the top: <input type='text' id='e_top_id' class='form_start' maxlength='40' size='15' value='familiar to you?'> <br> <br>Caption at the left: <input type='text' id='e_left_id' class='form_start' maxlength='40' size='15' value='unfamiliar = 'E''> <br> <br>Caption at the right: <input type='text' id='e_right_id' class='form_start' maxlength='40' size='15' value='familiar = 'I''> </div> <div style='flex: 50%;'> <b><u>Inducers in the enhanced version</u></b> <br> <
[…]
