+++ js/cit.js
@@ -110,6 +110,7 @@ function inducersAdded(groupOf6) {
         if (stim_162_base[indx].blur == "yes") {
             inducerIndex = the9fills.indexOf(stim_162_base[indx].filler);
             stim_162_base.splice(indx, 0, inducers[inducerIndex]);
+            stim_162_base.splice(indx, 0, inducers[inducerIndex]);
             indx++;
         }
     }
@@ -245,7 +246,7 @@ function randomDegrade(arrayOf6dicts) {
             temp6 = [groupOf6_rand[0]];
             fail = 0;
             var leng = groupOf6_rand.length; // actually 5 if there is no target
-            for (var i = 1; i < leng+1; i++) {
+            for (var i = 1; i < leng; i++) {
                 var newDictItem = groupOf6_rand[i];
                 var skip = [];
                 for (placeIndex = 0; placeIndex < temp6.length; placeIndex++) {
