### 2019-01-27 2f6744e
travis testing added
