+++ test/test.js
@@ -1,0 +1,5 @@
+describe("suite", function() {
+  it("check if dev_test() returns true", function() {
+    expect( dev_test() ).toBe(true);
+  });
+});
