### 2019-01-27 gasparl/citapp_pc#9 @gasparl
@andytwoods, @vsoch: I added the _dev\_test()_ (see my first comment) to be automatically run on each push [via Travis](https://travis-ci.org/gasparl/citapp_pc), using Karma/Jasmin/HeadlessChrome.

No libraries or any external files are included (or needed) in the app.

@vsoch: Nonetheless, if you are still curious whether your solution can also work, I can of course easily create a variable that changes when the test reached the end with success. So, if I understand what you need, I could e.g. assign a _window.test\_completed = false_ variable, that would become _true_ in the end.

Also, I think it wouldn't be hard to implement dev_test(1) and dev_test(2) in my current solution, but it doesn't seem very useful. All the main functions are covered with the simple dev_test(), the rest is just small HTML display manipulation, that is tricky to really check via pure coding.
