### 2019-01-27 gasparl/citapp_pc#9 @vsoch
Okay here is what I think is going on:

 - npm test goes to packages.json, which has a scripts section to define the test command:

```
"scripts": {
    "test": "karma start test/karma.conf.js --single-run"
  },
```
- karma start runs the file in test/karma.conf.js
- I see that files and the browser are loaded in this file [https://github.com/gasparl/citapp_pc/blob/master/test/karma.conf.js](https://github.com/gasparl/citapp_pc/blob/master/test/karma.conf.js) but what triggers some actual string of tests for the experiment? I would expect it to be longer than 0+0+0.076 seconds.
