### 2019-01-27 gasparl/citapp_pc#9 @gasparl
Yes. The Jasmine test itself is as simple as it can be:

```
describe("suite", function() {
  it("check if dev_test() returns true", function() {
    expect( dev_test() ).toBe(true);
  });
});
```

This is in test/test.js.

What "triggers" the test is the [_dev\_test()_](https://github.com/gasparl/citapp_pc/blob/master/js/citapp_dev_test.js) function here, which returns `true` if it passed through all main functions (preparing basic variables, CIT conditions, generating all items, initiating first block).

This does not involve any of the time-consuming parts, i.e. where participants have to respond to the items. That part is really quite simple, basically .show() and .hide() the items in the center on keypresses.

So all these tested functions are executed basically instantaneously.
