### 2019-01-27 gasparl/citapp_pc#9 @vsoch
ah understood - so the functions are triggered as soon as the browser loads I take it? We can open up discussion to add @andytwoods but would it hurt to add dev_test(1) or dev_test(2) since (when I tested the experiment, as much as I can remember) this seemed to actually go through the experiment (and thus would trigger errors if something was broken beyond loading). @andytwoods what are your thoughts?
