+++ test/test.js
@@ -1,6 +1,16 @@
-dev_test();
+
+function end_test() {
+    describe("suite", function() {
+
+      console.log(jasmine.DEFAULT_TIMEOUT_INTERVAL);
+      jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
 
 
+      it("All tests passed.", function() {
+        expect( true ).toBe(true);
+      });
+    });
+}
 function run_sims(sim_type) {
     dev_test(sim_type);
     count_checks = 0;
@@ -18,9 +28,7 @@ function run_sims(sim_type) {
                 run_sims(sim_type+1);
             } else {
                 console.log('Simulation 2 (enhanced) finished.');
-                it("All tests passed.", function() {
-                  expect( true ).toBe(true);
-                });
+                end_test();
             }
         } else if ( count_checks*10/60 > 7 ) {
             clearInterval(check_end);
@@ -29,7 +37,5 @@ function run_sims(sim_type) {
         }
     }, 1000*10 );
 }
-
-describe("suite", function() {
-  run_sims(1);
-});
+dev_test();
+run_sims(1);
