### 2019-01-27 gasparl/citapp_pc#9 @vsoch
Beautiful! [Here is the link on Travis](https://travis-ci.org/gasparl/citapp_pc) for @andytwoods. @gasparl you've definitely checked off all the boxes for my review, let's get this published soon! :)

If you like badges, note that you can grab the Travis Badge in the settings for the repository (on travis) and put in your README.md in case any future person (including yourself!) wants an easy link.
