### 2019-01-27 #1179 @vsoch
@gasparl has gone above and beyond to add [automated testing](https://travis-ci.org/gasparl/citapp_pc) for his application. I have checked this off as the final box to finish my review, and I'm ready to move forward with acceptance of the work for publication in JOSS.
