### 2019-01-27 gasparl/citapp_pc#9 @gasparl
Thanks again @vsoch. (I added the Travis Badge ;) )
