### 2019-01-27 b6d5e99
paper small corrections
