### 2019-01-31 f1fc562
probe irr mean sd display at end
