### 2019-01-31 e3a4294
probe irr mean sd display at end
