### 2019-02-09 #1179 @yochannah
Okay, I've run through the paper again and it all looks good - links in the references work and I didn't spot any typos in need of correction ✅

@gasparl - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? It's probably worth giving the manuscript a final proofread if you feel you need to, as well.
