### 2019-02-10 #1179 @gasparl
By the way, the version is now 1.0.3, not 1.0.0 (as on this page). Also, the Repository Address (and name) of the submitted paper (http://joss.theoj.org/papers/b2d9ee93da00cb5b3fdc962d50df5063) is still the old bitbucket link.
