### 2019-02-10 #1179 @yochannah
@gasparl
1. I've updated the version to be 1.0.3 - Thanks!
2. The proof above has the newer GitHub repo linked, and I noticed that one of the reviewers links still has BitBucket so I replaced it with the github link. If this all looks okay to you (I think it should be fine) I'll ping the editors-in-chief team and we'll accept the paper! 🎉
