### 2019-02-10 #1179 @gasparl
Though, just to be sure, regarding the old repository address, what I meant is that at http://joss.theoj.org/papers/active?page=7, I see this:
![image](https://user-images.githubusercontent.com/37374850/52535602-b93e1d00-2d50-11e9-9ae2-177f1086a65d.png)
And then also when I click on it:
![image](https://user-images.githubusercontent.com/37374850/52535610-cb1fc000-2d50-11e9-8ffe-34991911cba5.png)

But then perhaps this will automatically change when accepted?
