### 2019-02-10 #1179 @yochannah
hmmm, there's definitely nothing in the whedon commands listed there to update the repo URL.

@openjournals/joss-eics - Can you help? Somewhere some metadata still has the old repo pointing to it though - see [this comment](https://github.com/openjournals/joss-reviews/issues/1179#issuecomment-462141982) from @gasparl above. I've modified any links in the pre-review and review issues to the correct URL, but maybe I missed something?

- **correct repo:** https://github.com/gasparl/citapp_pc
- **wrong repo:** https://bitbucket.org/gaspar_lukacs/citapp_pc

**Other than that, I think we're ready to accept this paper!!** 🎉 💯 🏆 Thank you so much @vsoch and @andytwoods for your reviews and @gasparl for working with the reviewers to get this repo ready for publication.
