### 2019-02-12 #1179 @arfon
> @openjournals/joss-eics - Can you help? Somewhere some metadata still has the old repo pointing to it though - see [this comment](https://github.com/openjournals/joss-reviews/issues/1179#issuecomment-462141982) from @gasparl above. I've modified any links in the pre-review and review issues to the correct URL, but maybe I missed something?

Sorry, this is a manual process to update on the JOSS site. I've done this now.
